+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.0] - 2024-09-24
+
+### Added
+
+- Generated Delaunay Triangulation Graph for dungeon paths.
+  - Ensures paths are connected to each neighbouring room.
+
 ## [0.1.0] - 2024-09-20
 
 ### Added